refactor: Move network infra and utils to :ychat-core
