+++ ychat-core/src/commonMain/kotlin/co/yml/ychat/core/model/FileBytes.kt
@@ -1,0 +1,5 @@
+package co.yml.ychat.core.model
+
+expect class FileBytes
+
+expect fun FileBytes.toByteArray(): ByteArray
