+++ ychat-core/src/commonMain/kotlin/co/yml/ychat/core/network/factories/HttpClientFactory.kt
@@ -1,0 +1,7 @@
+package co.yml.ychat.core.network.factories
+
+import io.ktor.client.HttpClient
+
+interface HttpClientFactory {
+    fun getHttpClient(): HttpClient
+}
