+++ ychat-core/src/commonMain/kotlin/co/yml/ychat/core/network/factories/HttpEngineFactory.kt
@@ -1,0 +1,7 @@
+package co.yml.ychat.core.network.factories
+
+import io.ktor.client.engine.HttpClientEngine
+
+expect object HttpEngineFactory {
+    actual fun getEngine(): HttpClientEngine
+}
