+++ ychat-core/src/jvmMain/kotlin/co/yml/ychat/core/model/FileBytes.kt
@@ -1,0 +1,7 @@
+package co.yml.ychat.core.model
+
+actual typealias FileBytes = ByteArray
+
+actual fun FileBytes.toByteArray(): ByteArray {
+    return this
+}
