+++ ychat-core/src/jvmMain/kotlin/co/yml/ychat/core/network/factories/HttpEngineFactory.kt
@@ -1,0 +1,11 @@
+package co.yml.ychat.core.network.factories
+
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.okhttp.OkHttp
+
+actual object HttpEngineFactory {
+
+    actual fun getEngine(): HttpClientEngine {
+        return OkHttp.create()
+    }
+}
