+++ ychat/src/commonMain/kotlin/co/yml/ychat/domain/mapper/AudioMapper.kt
@@ -1,8 +1,8 @@
 package co.yml.ychat.domain.mapper
 
+import co.yml.ychat.core.model.FileBytes
 import co.yml.ychat.data.dto.AudioParamsDto
 import co.yml.ychat.domain.model.AudioParams
-import co.yml.ychat.domain.model.FileBytes
 
 internal fun AudioParams.toAudioParamsDto(filename: String, fileBytes: FileBytes): AudioParamsDto {
     return AudioParamsDto(
