--- ychat/src/commonMain/kotlin/co/yml/ychat/domain/model/FileBytes.kt
@@ -1,5 +1,0 @@
-package co.yml.ychat.domain.model
-
-expect class FileBytes
-
-internal expect fun FileBytes.toByteArray(): ByteArray
