+++ ychat/src/commonMain/kotlin/co/yml/ychat/domain/usecases/AudioUseCase.kt
@@ -1,9 +1,9 @@
 package co.yml.ychat.domain.usecases
 
+import co.yml.ychat.core.model.FileBytes
 import co.yml.ychat.data.api.ChatGptApi
 import co.yml.ychat.domain.mapper.toAudioParamsDto
 import co.yml.ychat.domain.model.AudioParams
-import co.yml.ychat.domain.model.FileBytes
 
 internal class AudioUseCase(private val chatGptApi: ChatGptApi) {
