+++ ychat/src/commonMain/kotlin/co/yml/ychat/entrypoint/features/RetrieveModel.kt
@@ -1,7 +1,7 @@
 package co.yml.ychat.entrypoint.features
 
 import co.yml.ychat.YChat
-import co.yml.ychat.data.exception.ChatGptException
+import co.yml.ychat.core.exceptions.YChatException
 import co.yml.ychat.domain.model.AIModel
 import kotlin.coroutines.cancellation.CancellationException
 
@@ -14,9 +14,9 @@ interface RetrieveModel {
      * @param id The ID of the model to retrieve.
      * @return The artificial intelligence model with the given [id].
      * @throws CancellationException if the operation is cancelled.
-     * @throws ChatGptException if there is an error generating chat completions.
+     * @throws YChatException if there is an error generating chat completions.
      */
-    @Throws(CancellationException::class, ChatGptException::class)
+    @Throws(CancellationException::class, YChatException::class)
     suspend fun execute(id: String): AIModel
 
     /**
