--- ychat/src/iosMain/kotlin/co/yml/ychat/domain/model/FileBytes.kt
@@ -1,16 +1,0 @@
-package co.yml.ychat.domain.model
-
-import kotlinx.cinterop.addressOf
-import kotlinx.cinterop.usePinned
-import platform.Foundation.NSData
-import platform.posix.memcpy
-
-actual typealias FileBytes = NSData
-
-actual fun FileBytes.toByteArray(): ByteArray {
-    return ByteArray(this.length.toInt()).apply {
-        usePinned {
-            memcpy(it.addressOf(0), this@toByteArray.bytes, this@toByteArray.length)
-        }
-    }
-}
