--- ychat/src/jvmMain/kotlin/co/yml/ychat/di/module/platformModule.kt
@@ -1,8 +1,0 @@
-package co.yml.ychat.di.module
-
-import io.ktor.client.engine.okhttp.OkHttp
-import org.koin.dsl.module
-
-internal actual fun platformModule() = module {
-    single { OkHttp.create() }
-}
