--- ychat/src/jvmMain/kotlin/co/yml/ychat/domain/model/FileBytes.kt
@@ -1,7 +1,0 @@
-package co.yml.ychat.domain.model
-
-actual typealias FileBytes = ByteArray
-
-internal actual fun FileBytes.toByteArray(): ByteArray {
-    return this
-}
